% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_lifecycle_hooks}
\alias{autoscaling_describe_lifecycle_hooks}
\title{Describes the lifecycle hooks for the specified Auto Scaling group}
\usage{
autoscaling_describe_lifecycle_hooks(AutoScalingGroupName,
  LifecycleHookNames)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{LifecycleHookNames}{The names of one or more lifecycle hooks. If you omit this parameter,
all lifecycle hooks are described.}
}
\description{
Describes the lifecycle hooks for the specified Auto Scaling group.
}
\section{Request syntax}{
\preformatted{svc$describe_lifecycle_hooks(
  AutoScalingGroupName = "string",
  LifecycleHookNames = list(
    "string"
  )
)
}
}

\examples{
# This example describes the lifecycle hooks for the specified Auto
# Scaling group.
\donttest{svc$describe_lifecycle_hooks(
  AutoScalingGroupName = "my-auto-scaling-group"
)}

}
\keyword{internal}
