% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_termination_policy_types}
\alias{autoscaling_describe_termination_policy_types}
\title{Describes the termination policies supported by Amazon EC2 Auto Scaling}
\usage{
autoscaling_describe_termination_policy_types()
}
\description{
Describes the termination policies supported by Amazon EC2 Auto Scaling.
}
\details{
For more information, see \href{https://docs.aws.amazon.com/autoscaling/ec2/userguide/as-instance-termination.html}{Controlling Which Auto Scaling Instances Terminate During Scale In}
in the \emph{Amazon EC2 Auto Scaling User Guide}.
}
\examples{
# This example describes the available termination policy types.
\donttest{svc$describe_termination_policy_types()}

}
\keyword{internal}
