% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_delete_change_set}
\alias{cloudformation_delete_change_set}
\title{Deletes the specified change set}
\usage{
cloudformation_delete_change_set(ChangeSetName, StackName)
}
\arguments{
\item{ChangeSetName}{[required] The name or Amazon Resource Name (ARN) of the change set that you want
to delete.}

\item{StackName}{If you specified the name of a change set to delete, specify the stack
name or ID (ARN) that is associated with it.}
}
\description{
Deletes the specified change set. Deleting change sets ensures that no
one executes the wrong change set.
}
\details{
If the call successfully completes, AWS CloudFormation successfully
deleted the change set.
}
\section{Request syntax}{
\preformatted{svc$delete_change_set(
  ChangeSetName = "string",
  StackName = "string"
)
}
}

\keyword{internal}
