% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_describe_stack_set_operation}
\alias{cloudformation_describe_stack_set_operation}
\title{Returns the description of the specified stack set operation}
\usage{
cloudformation_describe_stack_set_operation(StackSetName, OperationId)
}
\arguments{
\item{StackSetName}{[required] The name or the unique stack ID of the stack set for the stack
operation.}

\item{OperationId}{[required] The unique ID of the stack set operation.}
}
\description{
Returns the description of the specified stack set operation.
}
\section{Request syntax}{
\preformatted{svc$describe_stack_set_operation(
  StackSetName = "string",
  OperationId = "string"
)
}
}

\keyword{internal}
