% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_put_resource_policy}
\alias{cloudwatchlogs_put_resource_policy}
\title{Creates or updates a resource policy allowing other AWS services to put
log events to this account, such as Amazon Route 53}
\usage{
cloudwatchlogs_put_resource_policy(policyName, policyDocument)
}
\arguments{
\item{policyName}{Name of the new policy. This parameter is required.}

\item{policyDocument}{Details of the new policy, including the identity of the principal that
is enabled to put logs to this account. This is formatted as a JSON
string.

The following example creates a resource policy enabling the Route 53
service to put DNS query logs in to the specified log group. Replace
"logArn" with the ARN of your CloudWatch Logs resource, such as a log
group or log stream.

\code{\{ "Version": "2012-10-17", "Statement": \\[ \{ "Sid": "Route53LogsToCloudWatchLogs", "Effect": "Allow", "Principal": \{ "Service": \\[ "route53.amazonaws.com" \\] \}, "Action":"logs:PutLogEvents", "Resource": "logArn" \} \\] \} }}
}
\description{
Creates or updates a resource policy allowing other AWS services to put
log events to this account, such as Amazon Route 53. An account can have
up to 10 resource policies per region.
}
\section{Request syntax}{
\preformatted{svc$put_resource_policy(
  policyName = "string",
  policyDocument = "string"
)
}
}

\keyword{internal}
