% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_delivery_channel}
\alias{configservice_delete_delivery_channel}
\title{Deletes the delivery channel}
\usage{
configservice_delete_delivery_channel(DeliveryChannelName)
}
\arguments{
\item{DeliveryChannelName}{[required] The name of the delivery channel to delete.}
}
\description{
Deletes the delivery channel.
}
\details{
Before you can delete the delivery channel, you must stop the
configuration recorder by using the StopConfigurationRecorder action.
}
\section{Request syntax}{
\preformatted{svc$delete_delivery_channel(
  DeliveryChannelName = "string"
)
}
}

\keyword{internal}
