% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_retention_configuration}
\alias{configservice_delete_retention_configuration}
\title{Deletes the retention configuration}
\usage{
configservice_delete_retention_configuration(RetentionConfigurationName)
}
\arguments{
\item{RetentionConfigurationName}{[required] The name of the retention configuration to delete.}
}
\description{
Deletes the retention configuration.
}
\section{Request syntax}{
\preformatted{configservice$delete_retention_configuration(
  RetentionConfigurationName = "string"
)
}
}

\keyword{internal}
