% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_configuration_aggregators}
\alias{configservice_describe_configuration_aggregators}
\title{Returns the details of one or more configuration aggregators}
\usage{
configservice_describe_configuration_aggregators(
  ConfigurationAggregatorNames, NextToken, Limit)
}
\arguments{
\item{ConfigurationAggregatorNames}{The name of the configuration aggregators.}

\item{NextToken}{The nextToken string returned on a previous page that you use to get the
next page of results in a paginated response.}

\item{Limit}{The maximum number of configuration aggregators returned on each page.
The default is maximum. If you specify 0, AWS Config uses the default.}
}
\description{
Returns the details of one or more configuration aggregators. If the
configuration aggregator is not specified, this action returns the
details for all the configuration aggregators associated with the
account.
}
\section{Request syntax}{
\preformatted{configservice$describe_configuration_aggregators(
  ConfigurationAggregatorNames = list(
    "string"
  ),
  NextToken = "string",
  Limit = 123
)
}
}

\keyword{internal}
