% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_remediation_configurations}
\alias{configservice_describe_remediation_configurations}
\title{Returns the details of one or more remediation configurations}
\usage{
configservice_describe_remediation_configurations(ConfigRuleNames)
}
\arguments{
\item{ConfigRuleNames}{[required] A list of AWS Config rule names of remediation configurations for which
you want details.}
}
\description{
Returns the details of one or more remediation configurations.
}
\section{Request syntax}{
\preformatted{svc$describe_remediation_configurations(
  ConfigRuleNames = list(
    "string"
  )
)
}
}

\keyword{internal}
