% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_resource_inventory}
\alias{licensemanager_list_resource_inventory}
\title{Returns a detailed list of resources}
\usage{
licensemanager_list_resource_inventory(MaxResults, NextToken, Filters)
}
\arguments{
\item{MaxResults}{Maximum number of results to return in a single call. To retrieve the
remaining results, make another call with the returned \code{NextToken}
value.}

\item{NextToken}{Token for the next set of results.}

\item{Filters}{One or more filters.}
}
\description{
Returns a detailed list of resources.
}
\section{Request syntax}{
\preformatted{licensemanager$list_resource_inventory(
  MaxResults = 123,
  NextToken = "string",
  Filters = list(
    list(
      Name = "string",
      Condition = "EQUALS"|"NOT_EQUALS"|"BEGINS_WITH"|"CONTAINS",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
