% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworkscm_operations.R
\name{opsworkscm_describe_account_attributes}
\alias{opsworkscm_describe_account_attributes}
\title{Describes your account attributes, and creates requests to increase
limits before they are reached or exceeded}
\usage{
opsworkscm_describe_account_attributes()
}
\description{
Describes your account attributes, and creates requests to increase
limits before they are reached or exceeded.
}
\details{
This operation is synchronous.
}
\section{Request syntax}{
\preformatted{svc$describe_account_attributes()
}
}

\keyword{internal}
