% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_get_group_query}
\alias{resourcegroups_get_group_query}
\title{Returns the resource query associated with the specified resource group}
\usage{
resourcegroups_get_group_query(GroupName)
}
\arguments{
\item{GroupName}{[required] The name of the resource group.}
}
\description{
Returns the resource query associated with the specified resource group.
}
\section{Request syntax}{
\preformatted{svc$get_group_query(
  GroupName = "string"
)
}
}

\keyword{internal}
