% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_untag}
\alias{resourcegroups_untag}
\title{Deletes specified tags from a specified resource}
\usage{
resourcegroups_untag(Arn, Keys)
}
\arguments{
\item{Arn}{[required] The ARN of the resource from which to remove tags.}

\item{Keys}{[required] The keys of the tags to be removed.}
}
\description{
Deletes specified tags from a specified resource.
}
\section{Request syntax}{
\preformatted{svc$untag(
  Arn = "string",
  Keys = list(
    "string"
  )
)
}
}

\keyword{internal}
