% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_create_tag_option}
\alias{servicecatalog_create_tag_option}
\title{Creates a TagOption}
\usage{
servicecatalog_create_tag_option(Key, Value)
}
\arguments{
\item{Key}{[required] The TagOption key.}

\item{Value}{[required] The TagOption value.}
}
\description{
Creates a TagOption.
}
\section{Request syntax}{
\preformatted{servicecatalog$create_tag_option(
  Key = "string",
  Value = "string"
)
}
}

\keyword{internal}
