% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_product_view}
\alias{servicecatalog_describe_product_view}
\title{Gets information about the specified product}
\usage{
servicecatalog_describe_product_view(AcceptLanguage, Id)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The product view identifier.}
}
\description{
Gets information about the specified product.
}
\section{Request syntax}{
\preformatted{svc$describe_product_view(
  AcceptLanguage = "string",
  Id = "string"
)
}
}

\keyword{internal}
