% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_portfolios}
\alias{servicecatalog_list_portfolios}
\title{Lists all portfolios in the catalog}
\usage{
servicecatalog_list_portfolios(AcceptLanguage, PageToken, PageSize)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}

\item{PageSize}{The maximum number of items to return with this call.}
}
\description{
Lists all portfolios in the catalog.
}
\section{Request syntax}{
\preformatted{svc$list_portfolios(
  AcceptLanguage = "string",
  PageToken = "string",
  PageSize = 123
)
}
}

\keyword{internal}
