% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_cancel_maintenance_window_execution}
\alias{ssm_cancel_maintenance_window_execution}
\title{Stops a Maintenance Window execution that is already in progress and
cancels any tasks in the window that have not already starting running}
\usage{
ssm_cancel_maintenance_window_execution(WindowExecutionId)
}
\arguments{
\item{WindowExecutionId}{[required] The ID of the Maintenance Window execution to stop.}
}
\description{
Stops a Maintenance Window execution that is already in progress and
cancels any tasks in the window that have not already starting running.
(Tasks already in progress will continue to completion.)
}
\section{Request syntax}{
\preformatted{svc$cancel_maintenance_window_execution(
  WindowExecutionId = "string"
)
}
}

\keyword{internal}
