% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_automation_executions}
\alias{ssm_describe_automation_executions}
\title{Provides details about all active and terminated Automation executions}
\usage{
ssm_describe_automation_executions(Filters, MaxResults, NextToken)
}
\arguments{
\item{Filters}{Filters used to limit the scope of executions that are requested.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Provides details about all active and terminated Automation executions.
}
\section{Request syntax}{
\preformatted{svc$describe_automation_executions(
  Filters = list(
    list(
      Key = "DocumentNamePrefix"|"ExecutionStatus"|"ExecutionId"|"ParentExecutionId"|"CurrentAction"|"StartTimeBefore"|"StartTimeAfter"|"AutomationType",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
