% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_maintenance_window_execution}
\alias{ssm_get_maintenance_window_execution}
\title{Retrieves details about a specific task executed as part of a
Maintenance Window execution}
\usage{
ssm_get_maintenance_window_execution(WindowExecutionId)
}
\arguments{
\item{WindowExecutionId}{[required] The ID of the Maintenance Window execution that includes the task.}
}
\description{
Retrieves details about a specific task executed as part of a
Maintenance Window execution.
}
\section{Request syntax}{
\preformatted{svc$get_maintenance_window_execution(
  WindowExecutionId = "string"
)
}
}

\keyword{internal}
