% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_maintenance_window_execution_task}
\alias{ssm_get_maintenance_window_execution_task}
\title{Retrieves the details about a specific task executed as part of a
Maintenance Window execution}
\usage{
ssm_get_maintenance_window_execution_task(WindowExecutionId, TaskId)
}
\arguments{
\item{WindowExecutionId}{[required] The ID of the Maintenance Window execution that includes the task.}

\item{TaskId}{[required] The ID of the specific task execution in the Maintenance Window task
that should be retrieved.}
}
\description{
Retrieves the details about a specific task executed as part of a
Maintenance Window execution.
}
\section{Request syntax}{
\preformatted{svc$get_maintenance_window_execution_task(
  WindowExecutionId = "string",
  TaskId = "string"
)
}
}

\keyword{internal}
