% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_parameter_history}
\alias{ssm_get_parameter_history}
\title{Query a list of all parameters used by the AWS account}
\usage{
ssm_get_parameter_history(Name, WithDecryption, MaxResults, NextToken)
}
\arguments{
\item{Name}{[required] The name of a parameter you want to query.}

\item{WithDecryption}{Return decrypted values for secure string parameters. This flag is
ignored for String and StringList parameter types.}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Query a list of all parameters used by the AWS account.
}
\section{Request syntax}{
\preformatted{svc$get_parameter_history(
  Name = "string",
  WithDecryption = TRUE|FALSE,
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
