% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_describe_attachment}
\alias{support_describe_attachment}
\title{Returns the attachment that has the specified ID}
\usage{
support_describe_attachment(attachmentId)
}
\arguments{
\item{attachmentId}{[required] The ID of the attachment to return. Attachment IDs are returned by the
DescribeCommunications operation.}
}
\description{
Returns the attachment that has the specified ID. Attachment IDs are
generated by the case management system when you add an attachment to a
case or case communication. Attachment IDs are returned in the
AttachmentDetails objects that are returned by the
DescribeCommunications operation.
}
\section{Request syntax}{
\preformatted{svc$describe_attachment(
  attachmentId = "string"
)
}
}

\keyword{internal}
